+++ pytsite/image/_model.py
@@ -99,9 +99,9 @@ class Image(_file.model.File):
             }, strip_lang=True)
 
         elif field_name == 'thumb_url':
-            return self.f_get('url', width=int(kwargs.get('width', 422)), height=int(kwargs.get('height', 422)))
-
-        return super()._on_f_get(field_name, value, **kwargs)
+            return self.f_get('url', width=int(kwargs.get('width', 450)), height=int(kwargs.get('height', 450)))
+        else:
+            return super()._on_f_get(field_name, value, **kwargs)
 
     def get_url(self, width: int=0, height: int=0) -> str:
         """Shortcut to use in Jinja templates.